% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apocryphabooks.R
\docType{data}
\name{apocryphabooks}
\alias{apocryphabooks}
\title{Apocrypha Books.}
\format{A data frame with 31102 rows and 5 variables:
\describe{
  \item{Book.Abbreviation}{Abbreviation of book title}
  \item{King.James.Bible}{Book title in King James Bible}
  \item{Vulgate}{Book title in Vulgate}
  \item{Douay.Rheims}{Book title in Douay Rheims version}
  \item{Full.Title.Auth.V}{Book title in authorized version}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
apocryphabooks
}
\description{
A dataset containing all books of the Apocrypha for reference.
}
\keyword{datasets}

