% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_of_mormon.R
\docType{data}
\name{book_of_mormon}
\alias{book_of_mormon}
\title{Book of Mormon (LDS Text)}
\format{A \link[tibble]{tibble} with 6,604 rows and 5 variables:
\describe{
  \item{volume_title}{ Title of volume}
  \item{book_title}{ Title of book}
  \item{chapter_number}{ Chapter number}
  \item{verse_number}{ Verse number}
  \item{scripture_text}{ Verse text}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
book_of_mormon
}
\description{
A dataset containing all verses of the Book of Mormon, includes all 15
books, one verse per row.
}
\keyword{datasets}
