% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctrine_and_covenants.R
\docType{data}
\name{doctrine_and_covenants}
\alias{doctrine_and_covenants}
\title{Doctrine & Covenants (LDS Text)}
\format{A \link[tibble]{tibble} with 3,654 rows and 5 variables:
\describe{
  \item{volume_title}{ Title of volume}
  \item{book_title}{ Title of book}
  \item{chapter_number}{ Chapter number}
  \item{verse_number}{ Verse number}
  \item{scripture_text}{ Verse text}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
doctrine_and_covenants
}
\description{
A dataset containing all verses of the Doctrine & Covenants, one verse per row.
}
\keyword{datasets}
