% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_new_testament.R
\docType{data}
\name{greek_new_testament}
\alias{greek_new_testament}
\title{The Greek New Testament}
\format{A \link[tibble]{tibble} with 7,958 rows and 4 variables:
\describe{
  \item{book}{ Book name}
  \item{pslam}{ Psalm}
  \item{verse}{ Verse}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
greek_new_testament
}
\description{
A dataset containing all verses of the Greek New Testament, includes all 27
books, one verse per row.
}
\keyword{datasets}
