% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king_james_books.R
\docType{data}
\name{king_james_books}
\alias{king_james_books}
\title{King James Bible Books.}
\format{A data frame with 31102 rows and 7 variables:
\describe{
  \item{Book.Abbreviation}{Abbreviation of book title}
  \item{King.James.Bible}{Book title in King James Bible}
  \item{Vulgate}{Book title in Vulgate}
  \item{Douay.Rheims}{Book title in Douay Rheims version}
  \item{Full.Title.Auth.V}{Book title in authorized version}
  \item{Testament}{Testament to which book belongs}
  \item{Book.Number}{Book sequence as appears in King James Bible}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
king_james_books
}
\description{
A dataset containing all books of the King James Bible for reference.
}
\keyword{datasets}

