% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kingjamesbooks.R
\docType{data}
\name{kingjamesbooks}
\alias{kingjamesbooks}
\title{King James Bible Books.}
\format{A data frame with 31102 rows and 6 variables:
\describe{
  \item{Book.Abbreviation}{Abbreviation of book title}
  \item{King.James.Bible}{Book title in King James Bible}
  \item{Vulgate}{Book title in Vulgate}
  \item{Douay.Rheims}{Book title in Douay Rheims version}
  \item{Full.Title.Auth.V}{Book title in authorized version}
  \item{Testament}{Testament to which book belongs}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
kingjamesbooks
}
\description{
A dataset containing all books of the King James Bible for reference.
}
\keyword{datasets}

