% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_terms.R
\docType{data}
\name{middle_english_negative}
\alias{middle_english_negative}
\title{Middle English negative terms}
\format{a character vector of 41 negative words}
\usage{
middle_english_negative
}
\description{
A character vector of Middle English negative terms suitable for sentiment
analysis.
}
\keyword{datasets}

