% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_english_sentiments.R
\docType{data}
\name{middle_english_sentiments}
\alias{middle_english_sentiments}
\title{Middle English sentiments.}
\format{A \link[tibble]{tibble} with 90 rows and 3 variables:
\describe{
  \item{word}{ term}
  \item{sentiment}{ positive or negative}
  \item{score}{ 1 for positive and -1 for negative}
}}
\usage{
middle_english_sentiments
}
\description{
A dataset containing Middle English positive and negative terms.
}
\keyword{datasets}
