% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{middle_english_stopwords}
\alias{middle_english_stopwords}
\title{Middle English stopwords.}
\format{A character vector of 88 term.}
\usage{
middle_english_stopwords
}
\description{
A character vector of 92 Middle English stopwords.
}
\keyword{datasets}

