% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearl_of_great_price.R
\docType{data}
\name{pearl_of_great_price}
\alias{pearl_of_great_price}
\title{Pearl of Great Price (LDS Text)}
\format{A \link[tibble]{tibble} with 635 rows and 5 variables:
\describe{
  \item{volume_title}{ Title of volume}
  \item{book_title}{ Title of book}
  \item{chapter_number}{ Chapter number}
  \item{verse_number}{ Verse number}
  \item{scripture_text}{ Verse text}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
pearl_of_great_price
}
\description{
A dataset containing all verses of the Pearl of Great Price, one verse per row.
}
\keyword{datasets}
