% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_terms.R
\docType{data}
\name{positive_terms}
\alias{positive_terms}
\title{Middle English positive terms}
\format{a character vector of 48 positive words}
\usage{
positive_terms
}
\description{
A character vector of Middle English positive terms suitable for sentiment
analysis.
}
\keyword{datasets}

