% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacred.R
\docType{package}
\name{sacred}
\alias{sacred}
\alias{sacred-package}
\title{Sacred texts in tidy formats}
\description{
Sacred texts in tidy formats
}
\section{Datasets}{

\itemize{
  \item{\code{king_james_version}}{ \link[tibble]{tibble} of the King James bible, one row per verse.}
  \item{\code{apocrypha}}{ \link[tibble]{tibble} of books in the Apocrypha, one row per verse.}
  \item{\code{vulgate}}{ \link[tibble]{tibble} of books in the Vulgate, one row per verse.}
  \item{\code{tanach}}{ \link[tibble]{tibble} of books in the Tanach, one row per verse.}
  \item{\code{septuagint}}{ \link[tibble]{tibble} of books in the Septuagint, one row per verse.}
  \item{\code{greek_new_testament}}{ \link[tibble]{tibble} of books in the Greek New Testatement, one row per verse.}
  \item{\code{middle_english_stopwords}}{ \link[tibble]{tibble} of Middle English stopwords.}
  \item{\code{middle_english_sentiments}}{ \link[tibble]{tibble} lexicon of Middle English positive and negative terms.}
}
}

