% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apocrypha.R
\docType{data}
\name{apocrypha}
\alias{apocrypha}
\title{Apocrypha dataframe.}
\format{A \link[tibble]{tibble} with 5,725 rows and 5 variables:
\describe{
  \item{book.num}{ Book number}
  \item{book}{ Book name}
  \item{psalm}{ Psalm}
  \item{verse}{ Verse}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
apocrypha
}
\description{
A dataset containing all verses of the Apocrypha, includes all 16
books, one verse per row.
}
\keyword{datasets}
