% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king_james_version.R
\docType{data}
\name{king_james_version}
\alias{king_james_version}
\title{King James Bible.}
\format{A \link[tibble]{tibble} with 31,102 rows and 4 variables:
\describe{
  \item{book}{ Book name}
  \item{psalm}{ Psalm}
  \item{verse}{ Verse}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
king_james_version
}
\description{
A dataset containing all verses of the King James Bible, includes all 66
books, one verse per row.
}
\keyword{datasets}
