% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{middle_english_stopwords}
\alias{middle_english_stopwords}
\title{Middle English stopwords.}
\format{A \link[tibble]{tibble} with 295 rows and 2 variables:
\describe{
  \item{word}{ term}
  \item{lexicon}{ Middle English}
}}
\source{
\href{https://github.com/cltk/cltk/blob/da1b58838005a76c7f7ce239b23b17f4447df1d2/cltk/stop/old_english/stops.py}{Classical Language Toolkit}
}
\usage{
middle_english_stopwords
}
\description{
A dataset containing Middle English stopwords.
}
\keyword{datasets}
