% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/septuagint.R
\docType{data}
\name{septuagint}
\alias{septuagint}
\title{The Septuagint.}
\format{A \link[tibble]{tibble} with 30,374 rows and 3 variables:
\describe{
  \item{verse}{ Verse}
  \item{book}{ Book name}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
septuagint
}
\description{
A dataset containing all verses of the Septuagint, includes all 59
books, one verse per row.
}
\keyword{datasets}
