% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanach.R
\docType{data}
\name{tanach}
\alias{tanach}
\title{The Tanach.}
\format{A \link[tibble]{tibble} with 23,213 rows and 4 variables:
\describe{
  \item{book}{ Book name}
  \item{pslam}{ Psalm}
  \item{verse}{ Verse}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
tanach
}
\description{
A dataset containing all verses of the Tanach, includes all 39
books, one verse per row.
}
\keyword{datasets}
