% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulgate.R
\docType{data}
\name{vulgate}
\alias{vulgate}
\title{The Vulgate.}
\format{A \link[tibble]{tibble} with 35,811 rows and 3 variables:
\describe{
  \item{verse}{ Verse}
  \item{book}{ Book name}
  \item{text}{ Text of verse}
}}
\source{
\url{http://www.sacred-texts.com/bib/osrc/}
}
\usage{
vulgate
}
\description{
A dataset containing all verses of the Vulgate, includes all 73
books, one verse per row.
}
\keyword{datasets}
